src/bot/requirements.txt
requests==2.29.0
python-dotenv==1.0.0
pydantic==1.10.7
websocket-client==1.5.1 